decouple MagicMethodDetector
